move oauth code to separate module
