+++ src/main.rs
@@ -10,13 +10,12 @@ use std::{
     collections::{BTreeMap, BTreeSet},
     env,
     net::{IpAddr, Ipv4Addr, SocketAddr},
-    ops::Deref,
     path::PathBuf,
     sync::Arc,
     time::Duration,
 };
 
-use crate::{data::transcript::read_transcript_file, lobby::clear_lobby_on_interval};
+use crate::{data::transcript::read_transcript_file, lobby::clear_lobby_on_interval, oauth::{siwe_oauth_client, github_oauth_client}};
 use axum::{
     extract::Extension,
     response::Html,
@@ -27,7 +26,7 @@ use chrono::{DateTime, FixedOffset};
 use clap::Parser;
 use cli_batteries::{await_shutdown, version};
 use eyre::{bail, ensure, eyre, Result as EyreResult};
-use oauth2::{basic::BasicClient, AuthUrl, ClientId, ClientSecret, RedirectUrl, TokenUrl};
+use oauth::{CsrfToken, IdTokenSub};
 use sessions::{SessionId, SessionInfo};
 use storage::persistent_storage_client;
 use tokio::sync::RwLock;
@@ -42,11 +41,7 @@ use crate::{
         info::{current_state, jwt_info, status},
         lobby::try_contribute,
     },
-    constants::{
-        GITHUB_OAUTH_AUTH_URL, GITHUB_OAUTH_REDIRECT_URL, GITHUB_OAUTH_TOKEN_URL,
-        LOBBY_FLUSH_INTERVAL,
-        SIWE_OAUTH_AUTH_URL, SIWE_OAUTH_REDIRECT_URL, SIWE_OAUTH_TOKEN_URL,
-    },
+    constants::LOBBY_FLUSH_INTERVAL,
     data::transcript::{Contribution, Transcript},
     keys::Keys,
     test_transcript::TestTranscript,
@@ -57,6 +52,7 @@ mod constants;
 mod data;
 mod jwt;
 mod keys;
+mod oauth;
 mod lobby;
 mod sessions;
 mod storage;
@@ -137,71 +133,6 @@ where
     Ok(())
 }
 
-#[derive(Clone)]
-pub struct SiweOAuthClient {
-    client: BasicClient,
-}
-
-impl Deref for SiweOAuthClient {
-    type Target = BasicClient;
-
-    fn deref(&self) -> &Self::Target {
-        &self.client
-    }
-}
-
-fn siwe_oauth_client() -> SiweOAuthClient {
-    let client_id = env::var("SIWE_CLIENT_ID").expect("Missing SIWE_CLIENT_ID!");
-    let client_secret = env::var("SIWE_CLIENT_SECRET").expect("Missing SIWE_CLIENT_SECRET!");
-
-    let redirect_url =
-        env::var("SIWE_REDIRECT_URL").unwrap_or_else(|_| SIWE_OAUTH_REDIRECT_URL.to_string());
-    let auth_url = env::var("SIWE_AUTH_URL").unwrap_or_else(|_| SIWE_OAUTH_AUTH_URL.to_string());
-    let token_url = env::var("SIWE_TOKEN_URL").unwrap_or_else(|_| SIWE_OAUTH_TOKEN_URL.to_string());
-
-    SiweOAuthClient {
-        client: BasicClient::new(
-            ClientId::new(client_id),
-            Some(ClientSecret::new(client_secret)),
-            AuthUrl::new(auth_url).unwrap(),
-            Some(TokenUrl::new(token_url).unwrap()),
-        )
-        .set_redirect_uri(RedirectUrl::new(redirect_url).unwrap()),
-    }
-}
-
-#[derive(Clone)]
-pub struct GithubOAuthClient {
-    client: BasicClient,
-}
-
-impl Deref for GithubOAuthClient {
-    type Target = BasicClient;
-
-    fn deref(&self) -> &Self::Target {
-        &self.client
-    }
-}
-
-fn github_oauth_client() -> GithubOAuthClient {
-    let client_id = env::var("GITHUB_CLIENT_ID").expect("Missing GITHUB_CLIENT_ID!");
-    let client_secret = env::var("GITHUB_CLIENT_SECRET").expect("Missing GITHUB_CLIENT_SECRET!");
-    let redirect_url =
-        env::var("GITHUB_REDIRECT_URL").unwrap_or_else(|_| GITHUB_OAUTH_REDIRECT_URL.to_string());
-    let auth_url =
-        env::var("GITHUB_AUTH_URL").unwrap_or_else(|_| GITHUB_OAUTH_AUTH_URL.to_string());
-    let token_url =
-        env::var("GITHUB_TOKEN_URL").unwrap_or_else(|_| GITHUB_OAUTH_TOKEN_URL.to_string());
-    GithubOAuthClient {
-        client: BasicClient::new(
-            ClientId::new(client_id),
-            Some(ClientSecret::new(client_secret)),
-            AuthUrl::new(auth_url).unwrap(),
-            Some(TokenUrl::new(token_url).unwrap()),
-        )
-        .set_redirect_uri(RedirectUrl::new(redirect_url).unwrap()),
-    }
-}
 
 #[allow(clippy::unused_async)] // Required for axum function signature
 async fn hello_world() -> Html<&'static str> {
@@ -236,9 +167,6 @@ impl Default for AppConfig {
         }
     }
 }
-
-type IdTokenSub = String;
-type CsrfToken = String;
 
 #[derive(Default)]
 pub struct AppState {
